fix(streaming): count clipping only for delivered samples, and widen the counter
Three findings from the Qodo pass on #818, all real, and they share one fix.

The clip state was published per TICK -- before the priming gate and before
AInSampleList_PushBack decides whether the sample is emitted at all. A frame
suppressed at priming or lost to a full queue was still counted, so
ClippedSamples could disagree with what was actually delivered. Worse at
priming specifically: the cache path can present a stale value there that looks
exactly like the bottom rail, so the suppressed frame was the one most likely
to be a false positive.

Publishing now happens in the delivery branch, beside
gStreamStats.totalSamplesStreamed++ and inside the same taskENTER_CRITICAL.
That resolves the other two findings at the same time:

- the two RMWs (gClippedSamples++, gClipChannelMask |=) no longer rely on a
  single-writer argument -- they sit in the same critical section as their
  neighbour, which is what the CLAUDE.md atomicity rule asks for;
- gClippedSamples becomes uint64_t, matching totalSamplesStreamed. That is not
  cosmetic symmetry: a fully railed channel makes clippedSamples EQUAL to
  totalSamplesStreamed (bench-measured: 1922 of 1922), so it inherits exactly
  the same range requirement, and uint32 would wrap inside a long log. The
  64-bit access is already covered -- the snapshot read is inside the stats
  critical section, ClearStats' reset is inside its own, and Streaming_Init
  runs single-threaded like the gTimerISRCalls reset beside it.

Also from the same pass: the streaming.h comment claimed the live signal "rides
the protobuf device_status word", which contradicts the SCPI_SyncQuesBits
comment saying streaming frames deliberately omit device_status. The header was
the wrong one; it now states where the bit actually is (info and discovery
messages) and that the live in-session surface is STAT:QUES:COND? bit 0, read
over a link not carrying the stream.

Both configs build clean at -O3 -Werror. Re-flashed and re-ran the companion
test: 7/7, midscale 0/1922, fullscale 1922/1922 mask=1, QUES 1 during / 0 after.

Entire-Checkpoint: 4dd64a0fec1b

+++ firmware/src/services/SCPI/SCPIInterface.c
@@ -3247,7 +3247,7 @@ scpi_result_t SCPI_GetStreamStats(scpi_t * context) {
      * as a measurement, so it is absent from every loss percentage below. The
      * mask is over sample-list SLOTS (bit j = the j-th enabled channel in this
      * session), matching validMask, not over board channel numbers. */
-    scpi_printf(context, "ClippedSamples=%u\r\n", (unsigned)s.clippedSamples);
+    scpi_printf(context, "ClippedSamples=%llu\r\n", (unsigned long long)s.clippedSamples);
     scpi_printf(context, "ClippedChannelMask=%u\r\n", (unsigned)s.clippedChannelMask);
     // #541 D-A diag: ticks where a T1 result was not ready (ARDY clear) at
     // the deferred task's direct read.  Expected 0; non-zero ticks emitted
